# docs/swagger/swagger.yml
openapi: 3.0.3
info:
  title: Redesign API
  description: Redesign API
  contact:
    name: Peter Jerold Leslie
  version: v1
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://re-trustportal-api-jeroldleslie-dev.pacenthink.co/
    description: Development
  - url: https://re-trustportal-api-jeroldleslie-qa.pacenthink.co/
    description: QA
  - url: https://re-trustportal-api-jeroldleslie-uat.pacenthink.co/
    description: UAT
  - url: https://re-trustportal-api-jeroldleslie-main.pacenthink.co/
    description: Production
...
paths:
...
  /companies/{company_id}/users/{user_id}/security/awareness/users:
    post:
      tags:
        - Security Awareness
      description: Upload data for Security Awareness Campaign Users
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CompanyIdPathParameter'
        - $ref: '#/components/parameters/UserIdPathParameter'
      requestBody:
        content:
          text/csv:
            schema:
              type: string
              format: binary
      responses:
        200:
          description: Users added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: number
                      message:
                        type: string
        400:
          $ref: '#/components/responses/default400'
        401:
          $ref: '#/components/responses/default401'
        403:
          $ref: '#/components/responses/default403'
        404:
          $ref: '#/components/responses/default404'
        500:
          $ref: '#/components/responses/default500'
    get:
      tags:
        - Security Awareness
      description: Download Security Campaign users as a CSV file
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CompanyIdPathParameter'
        - $ref: '#/components/parameters/UserIdPathParameter'
      responses:
        200:
          description: Document converted successfully
          produces:
            text/csv:
              schema:
                type: file
            application/octet-stream:
              schema:
                type: string
                format: binary
        400:
          $ref: '#/components/responses/default400'
        401:
          $ref: '#/components/responses/default401'
        403:
          $ref: '#/components/responses/default403'
        404:
          $ref: '#/components/responses/default404'
        500:
          $ref: '#/components/responses/default500'
...
components:
  responses:
    default400:
      description: Request input was incorrect
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: "some error message"
    default401:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: "some error message"
    default403:
      description: No permission
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: "some error message"
    default404:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: "some error message"
    default500:
      description: Business logic error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: number
                    example: 500
                  message:
                    type: string
                    example: "some error message"
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  parameters:
    CompanyIdPathParameter:
      in: path
      name: company_id
      description: Company UUID
      required: true
      schema:
        type: string
        format: uuid
        example: 9940c9d2-8ef0-42b9-8626-5be545d1e63d
    UserIdPathParameter:
      in: path
      name: user_id
      description: User UUID
      required: true
      schema:
        type: string
        format: uuid
        example: 64d1802e-1fa4-4732-b182-0719199108a8
...